refactor: remove unused imports and create util save image function
