+++ src/pages/SearchResults/SearchResults.js
@@ -1,24 +1,11 @@
 import InfiniteLoader from 'react-infinite-loader';
-import useLocalStorage from 'utils/useLocalStorage';
+import SaveImage from 'utils/SaveImage';
 import { Wrapper, PhotoList } from 'components';
 
 function SearchResults({ data, getPhotos }) {
-  const [photos, setPhotos] = useLocalStorage('savedImages', []);
-
-  const handleSave = (content) => {
-    const imageExists = photos.find((ph) => ph.id === content.id);
-    if (imageExists) {
-      const photosArr = photos.filter((ph) => ph.id !== content.id);
-      setPhotos(photosArr);
-    }
-    if (!imageExists) {
-      const photosArr = [...photos, content];
-      setPhotos(photosArr);
-    }
-  };
   return (
     <Wrapper>
-      <PhotoList photos={data} handleSave={handleSave} />
+      <PhotoList photos={data} handleSave={SaveImage} />
       <InfiniteLoader onVisited={() => getPhotos()} />
     </Wrapper>
   );
